driver code moved to tests
